+++ packages/allow-deny/allow-deny.js
@@ -2,8 +2,6 @@
 /// Remote methods and access control.
 ///
 
-const hasOwn = Object.prototype.hasOwnProperty;
-
 // Restrict default mutators on collection. allow() and deny() take the
 // same options:
 //
@@ -138,17 +136,31 @@ CollectionPrototype._defineMutationMethods = function(options) {
           // between arbitrary client-specified _id fields and merely
           // client-controlled-via-randomSeed fields.
           let generatedId = null;
-          if (method === "insert" && !hasOwn.call(args[0], '_id')) {
-            generatedId = self._makeNewID();
+          let isBulkInsert = false;
+          if (method === "insert") {
+            isBulkInsert = true;
+            generatedId = [];
+            args.forEach(doc => {
+              if (!doc.hasOwnProperty('_id')) {
+                generatedId.push(self._makeNewID());
+              }
+            });
           }
 
           if (this.isSimulation) {
             // In a client simulation, you can do any mutation (even with a
             // complex selector).
-            if (generatedId !== null)
-              args[0]._id = generatedId;
+            if (generatedId !== null) {
+              if (!isBulkInsert) {
+                args[0]._id = generatedId;
+              } else {
+                args.forEach((doc, index) => {
+                  doc._id = generatedId[index];
+                });
+              }
+            }
             return self._collection[method].apply(
-              self._collection, args);
+              self._collection, [args]);
           }
 
           // This is the server receiving a method call from the client.
@@ -169,11 +181,18 @@ CollectionPrototype._defineMutationMethods = function(options) {
             const validatedMethodName =
                   '_validated' + method.charAt(0).toUpperCase() + method.slice(1);
             args.unshift(this.userId);
-            method === 'insert' && args.push(generatedId);
+            method === 'insert' && args.concat(generatedId);
             return self[validatedMethodName].apply(self, args);
           } else if (self._isInsecure()) {
-            if (generatedId !== null)
-              args[0]._id = generatedId;
+            if (generatedId !== null) {
+              if (isBulkInsert) {
+                args.forEach((doc, index) => {
+                  doc._id = generatedId[index];
+                });
+              } else {
+                args[0]._id = generatedId;
+              }
+            }
             // In insecure mode, allow any mutation (with a simple selector).
             // XXX This is kind of bogus.  Instead of blindly passing whatever
             //     we get from the network to this function, we should actually
@@ -297,7 +316,7 @@ CollectionPrototype._validatedUpdate = function(
     const params = mutator[op];
     if (op.charAt(0) !== '$') {
       throw new Meteor.Error(403, noReplaceError);
-    } else if (!hasOwn.call(ALLOWED_UPDATE_OPERATIONS, op)) {
+    } else if (!ALLOWED_UPDATE_OPERATIONS.hasOwnProperty(op)) {
       throw new Meteor.Error(
         403, "Access denied. Operator " + op + " not allowed in a restricted collection.");
     } else {
@@ -438,7 +457,7 @@ CollectionPrototype._callMutatorMethod = function _callMutatorMethod(name, args,
   const mutatorMethodName = this._prefix + name;
   return this._connection.apply(
     mutatorMethodName, args, { returnStubValue: true }, callback);
-}
+};
 
 function transformDoc(validator, doc) {
   if (validator.transform)
@@ -474,7 +493,7 @@ function addValidator(collection, allowOrDeny, options) {
   collection._restricted = true;
 
   ['insert', 'update', 'remove'].forEach((name) => {
-    if (hasOwn.call(options, name)) {
+    if (options.hasOwnProperty(name)) {
       if (!(options[name] instanceof Function)) {
         throw new Error(allowOrDeny + ": Value for `" + name + "` must be a function");
       }
@@ -510,7 +529,7 @@ function throwIfSelectorIsNotId(selector, methodName) {
       403, "Not permitted. Untrusted code may only " + methodName +
         " documents by ID.");
   }
-};
+}
 
 // Determine if we are in a DDP method simulation
 function alreadyInSimulation() {