+++ packages/mongo/mongo_driver.js
@@ -9,8 +9,8 @@
 
 const path = require("path");
 
-var MongoDB = NpmModuleMongodb;
-var Future = Npm.require('fibers/future');
+const MongoDB = NpmModuleMongodb;
+const Future = Npm.require('fibers/future');
 import { DocFetcher } from "./doc_fetcher.js";
 
 MongoInternals = {};
@@ -34,10 +34,10 @@ const APP_FOLDER = 'app';
 
 // This is used to add or remove EJSON from the beginning of everything nested
 // inside an EJSON custom type. It should only be called on pure JSON!
-var replaceNames = function (filter, thing) {
+const replaceNames = function (filter, thing) {
   if (typeof thing === "object" && thing !== null) {
-    if (_.isArray(thing)) {
-      return _.map(thing, _.bind(replaceNames, null, filter));
+    if (Array.isArray(thing)) {
+      return thing.map(_.bind(replaceNames, null, filter));
     }
     var ret = {};
     _.each(thing, function (value, key) {
@@ -56,12 +56,12 @@ MongoDB.Timestamp.prototype.clone = function () {
   return this;
 };
 
-var makeMongoLegal = function (name) { return "EJSON" + name; };
-var unmakeMongoLegal = function (name) { return name.substr(5); };
+const makeMongoLegal = function (name) { return "EJSON" + name; };
+const unmakeMongoLegal = function (name) { return name.substr(5); };
 
-var replaceMongoAtomWithMeteor = function (document) {
+const replaceMongoAtomWithMeteor = function (document) {
   if (document instanceof MongoDB.Binary) {
-    var buffer = document.value(true);
+    const buffer = document.value(true);
     return new Uint8Array(buffer);
   }
   if (document instanceof MongoDB.ObjectID) {
@@ -83,7 +83,7 @@ var replaceMongoAtomWithMeteor = function (document) {
   return undefined;
 };
 
-var replaceMeteorAtomWithMongo = function (document) {
+const replaceMeteorAtomWithMongo = function (document) {
   if (EJSON.isBinary(document)) {
     // This does more copies than we'd like, but is necessary because
     // MongoDB.BSON only looks like it takes a Uint8Array (and doesn't actually
@@ -111,11 +111,15 @@ var replaceMeteorAtomWithMongo = function (document) {
   return undefined;
 };
 
-var replaceTypes = function (document, atomTransformer) {
-  if (typeof document !== 'object' || document === null)
+const replaceTypes = function (document, atomTransformer, isBulk) {
+  if (!isBulk) {
+    if (typeof document !== 'object' || document === null)
+      return document;
+  } else if (!Array.isArray(document)) {
     return document;
+  }
 
-  var replacedTopLevelAtom = atomTransformer(document);
+  const replacedTopLevelAtom = atomTransformer(document);
   if (replacedTopLevelAtom !== undefined)
     return replacedTopLevelAtom;
 
@@ -309,7 +313,7 @@ MongoConnection.prototype._createCappedCollection = function (
 // after the observer notifiers have added themselves to the write
 // fence), you should call 'committed()' on the object returned.
 MongoConnection.prototype._maybeBeginWrite = function () {
-  var fence = DDPServer._CurrentWriteFence.get();
+  const fence = DDPServer._CurrentWriteFence.get();
   if (fence) {
     return fence.beginWrite();
   } else {
@@ -342,7 +346,7 @@ MongoConnection.prototype._onFailover = function (callback) {
 // well-defined -- a write "has been made" if it's returned, and an
 // observer "has been notified" if its callback has returned.
 
-var writeCallback = function (write, refresh, callback) {
+const writeCallback = function (write, refresh, callback) {
   return function (err, result) {
     if (! err) {
       // XXX We don't have to run this on error, right?
@@ -358,7 +362,7 @@ var writeCallback = function (write, refresh, callback) {
       }
     }
     write.committed();
-    if (callback) {
+    if (callback && typeof callback === 'function') {
       callback(err, result);
     } else if (err) {
       throw err;
@@ -366,43 +370,55 @@ var writeCallback = function (write, refresh, callback) {
   };
 };
 
-var bindEnvironmentForWrite = function (callback) {
+const bindEnvironmentForWrite = function (callback) {
   return Meteor.bindEnvironment(callback, "Mongo write");
 };
 
-MongoConnection.prototype._insert = function (collection_name, document,
+MongoConnection.prototype._insert = function (collection_name, documents,
                                               callback) {
-  var self = this;
+  const self = this;
+  const isBulkInsert = Array.isArray(documents);
+  documents = isBulkInsert ? documents : [ documents ];
 
-  var sendError = function (e) {
+  const sendError = function (e) {
     if (callback)
       return callback(e);
     throw e;
   };
 
   if (collection_name === "___meteor_failure_test_collection") {
-    var e = new Error("Failure test");
+    const e = new Error("Failure test");
     e._expectedByTest = true;
     sendError(e);
     return;
   }
 
-  if (!(LocalCollection._isPlainObject(document) &&
-        !EJSON._isCustomType(document))) {
+  const complexDoc = function (document) {
+    return !LocalCollection._isPlainObject(document) ||
+      EJSON._isCustomType(document);
+  };
+  if (documents.some(complexDoc)) {
     sendError(new Error(
       "Only plain objects may be inserted into MongoDB"));
     return;
   }
 
-  var write = self._maybeBeginWrite();
-  var refresh = function () {
-    Meteor.refresh({collection: collection_name, id: document._id });
+  const write = self._maybeBeginWrite();
+  const refresh = function () {
+    documents.forEach(function (document) {
+      Meteor.refresh({collection: collection_name, id: document._id });
+    });
   };
   callback = bindEnvironmentForWrite(writeCallback(write, refresh, callback));
   try {
-    var collection = self.rawCollection(collection_name);
-    collection.insert(replaceTypes(document, replaceMeteorAtomWithMongo),
-                      {safe: true}, callback);
+    const collection = self.rawCollection(collection_name);
+    if (isBulkInsert) {
+      collection.insertMany(replaceTypes(documents, replaceMeteorAtomWithMongo, isBulkInsert),
+        {safe: true}, callback);
+    } else {
+      collection.insertOne(replaceTypes(documents[0], replaceMeteorAtomWithMongo),
+        {safe: true}, callback);
+    }
   } catch (err) {
     write.committed();
     throw err;
@@ -412,14 +428,14 @@ MongoConnection.prototype._insert = function (collection_name, document,
 // Cause queries that may be affected by the selector to poll in this write
 // fence.
 MongoConnection.prototype._refresh = function (collectionName, selector) {
-  var refreshKey = {collection: collectionName};
+  const refreshKey = {collection: collectionName};
   // If we know which documents we're removing, don't poll queries that are
   // specific to other documents. (Note that multiple notifications here should
   // not cause multiple polls, since all our listener is doing is enqueueing a
   // poll.)
-  var specificIds = LocalCollection._idsMatchedBySelector(selector);
+  const specificIds = LocalCollection._idsMatchedBySelector(selector);
   if (specificIds) {
-    _.each(specificIds, function (id) {
+    specificIds.forEach(function (id) {
       Meteor.refresh(_.extend({id: id}, refreshKey));
     });
   } else {
@@ -432,7 +448,7 @@ MongoConnection.prototype._remove = function (collection_name, selector,
   var self = this;
 
   if (collection_name === "___meteor_failure_test_collection") {
-    var e = new Error("Failure test");
+    const e = new Error("Failure test");
     e._expectedByTest = true;
     if (callback) {
       return callback(e);
@@ -448,7 +464,7 @@ MongoConnection.prototype._remove = function (collection_name, selector,
   callback = bindEnvironmentForWrite(writeCallback(write, refresh, callback));
 
   try {
-    var collection = self.rawCollection(collection_name);
+    const collection = self.rawCollection(collection_name);
     var wrappedCallback = function(err, driverResult) {
       callback(err, transformResult(driverResult).numberAffected);
     };
@@ -471,7 +487,7 @@ MongoConnection.prototype._dropCollection = function (collectionName, cb) {
   cb = bindEnvironmentForWrite(writeCallback(write, refresh, cb));
 
   try {
-    var collection = self.rawCollection(collectionName);
+    const collection = self.rawCollection(collectionName);
     collection.drop(cb);
   } catch (e) {
     write.committed();
@@ -508,7 +524,7 @@ MongoConnection.prototype._update = function (collection_name, selector, mod,
   }
 
   if (collection_name === "___meteor_failure_test_collection") {
-    var e = new Error("Failure test");
+    const e = new Error("Failure test");
     e._expectedByTest = true;
     if (callback) {
       return callback(e);
@@ -661,7 +677,7 @@ MongoConnection.prototype._update = function (collection_name, selector, mod,
   }
 };
 
-var transformResult = function (driverResult) {
+const transformResult = function (driverResult) {
   var meteorResult = { numberAffected: 0 };
   if (driverResult) {
     var mongoResult = driverResult.result;
@@ -684,7 +700,7 @@ var transformResult = function (driverResult) {
 };
 
 
-var NUM_OPTIMISTIC_TRIES = 3;
+const NUM_OPTIMISTIC_TRIES = 3;
 
 // exposed for testing
 MongoConnection._isCannotChangeIdError = function (err) {
@@ -693,7 +709,7 @@ MongoConnection._isCannotChangeIdError = function (err) {
   // {name: String, code: Number, errmsg: String}
   // Older Mongo returns:
   // {name: String, code: Number, err: String}
-  var error = err.errmsg || err.err;
+  const error = err.errmsg || err.err;
 
   // We don't use the error code here
   // because the error code we observed it producing (16837) appears to be
@@ -781,7 +797,7 @@ var simulateUpsertWithInsertedId = function (collection, selector, mod,
   doUpdate();
 };
 
-_.each(["insert", "update", "remove", "dropCollection", "dropDatabase"], function (method) {
+["insert", "update", "remove", "dropCollection", "dropDatabase"].forEach((method) => {
   MongoConnection.prototype[method] = function (/* arguments */) {
     var self = this;
     return Meteor.wrapAsync(self["_" + method]).apply(self, arguments);
@@ -899,7 +915,7 @@ Cursor = function (mongo, cursorDescription) {
   self._synchronousCursor = null;
 };
 
-_.each(['forEach', 'map', 'fetch', 'count', Symbol.iterator], function (method) {
+['forEach', 'map', 'fetch', 'count', Symbol.iterator].forEach((method) => {
   Cursor.prototype[method] = function () {
     var self = this;
 
@@ -1385,7 +1401,7 @@ MongoConnection.prototype._observeChanges = function (
 // to InvalidationCrossbar.listen.
 
 listenAll = function (cursorDescription, listenCallback) {
-  var listeners = [];
+  const listeners = [];
   forEachTrigger(cursorDescription, function (trigger) {
     listeners.push(DDPServer._InvalidationCrossbar.listen(
       trigger, listenCallback));
@@ -1393,7 +1409,7 @@ listenAll = function (cursorDescription, listenCallback) {
 
   return {
     stop: function () {
-      _.each(listeners, function (listener) {
+      listeners.forEach((listener) => {
         listener.stop();
       });
     }
@@ -1458,7 +1474,7 @@ MongoConnection.prototype._observeChangesTailable = function (
   }
 
   return self.tail(cursorDescription, function (doc) {
-    var id = doc._id;
+    const id = doc._id;
     delete doc._id;
     // The ts is an implementation detail. Hide it.
     delete doc.ts;